Windows fixes.
- Use `UrlExt::to_uds_file_path` to parse management URI.

- Mount management socket directory instead of socket file, as required
  by Windows.

- Use `-C target-feature=+crt-static` to compile iotedge-diagnostics.exe
  so that it does not depend on MSVCRT DLL.

- Set "IOTEDGE_RUN_AS_CONSOLE" env var when querying iotedged version.

- Use new location of iotedged.exe under Program Files.
