+++ edgelet/edgelet-core/src/lib.rs
@@ -21,8 +21,14 @@ extern crate serde_derive;
 extern crate serde_json;
 extern crate sha2;
 extern crate tokio;
+extern crate url;
 
 extern crate edgelet_utils;
+
+use std::path::{Path, PathBuf};
+
+use failure::ResultExt;
+use url::Url;
 
 mod authorization;
 mod certificate_properties;
@@ -60,3 +66,44 @@ lazy_static! {
 pub fn version() -> &'static str {
     &VERSION
 }
+
+pub trait UrlExt {
+    fn to_uds_file_path(&self) -> Result<PathBuf, Error>;
+    fn to_base_path(&self) -> Result<PathBuf, Error>;
+}
+
+impl UrlExt for Url {
+    fn to_uds_file_path(&self) -> Result<PathBuf, Error> {
+        debug_assert_eq!(self.scheme(), UNIX_SCHEME);
+
+        if cfg!(windows) {
+            // We get better handling of Windows file syntax if we parse a
+            // unix:// URL as a file:// URL. Specifically:
+            // - On Unix, `Url::parse("unix:///path")?.to_file_path()` succeeds and
+            //   returns "/path".
+            // - On Windows, `Url::parse("unix:///C:/path")?.to_file_path()` fails
+            //   with Err(()).
+            // - On Windows, `Url::parse("file:///C:/path")?.to_file_path()` succeeds
+            //   and returns "C:\\path".
+            debug_assert_eq!(self.scheme(), UNIX_SCHEME);
+            let mut s = self.to_string();
+            s.replace_range(..4, "file");
+            let url = Url::parse(&s).with_context(|_| ErrorKind::InvalidUrl(s.clone()))?;
+            let path = url
+                .to_file_path()
+                .map_err(|()| ErrorKind::InvalidUrl(url.to_string()))?;
+            Ok(path)
+        } else {
+            Ok(Path::new(self.path()).to_path_buf())
+        }
+    }
+
+    fn to_base_path(&self) -> Result<PathBuf, Error> {
+        match self.scheme() {
+            "unix" => Ok(self.to_uds_file_path()?),
+            _ => Ok(self.as_str().into()),
+        }
+    }
+}
+
+pub const UNIX_SCHEME: &str = "unix";
