+++ edgelet/iotedge-diagnostics/Cargo.toml
@@ -3,6 +3,7 @@ name = "iotedge-diagnostics"
 version = "0.1.0"
 authors = ["Azure IoT Edge Devs"]
 publish = false
+edition = "2018"
 
 [dependencies]
 clap = "2.31"
@@ -12,6 +13,7 @@ serde_json = "1.0"
 tokio = "0.1"
 url = "1.7"
 
+edgelet-core = { path = "../edgelet-core" }
 management = { path = "../management" }
 
 [target.'cfg(unix)'.dependencies]
